+++ online-store/src/main/java/com/skwarek/onlineStore/data/model/product/specifications/NotebookSpecifications.java
@@ -1,13 +1,14 @@
-package com.skwarek.onlineStore.data.entity.product.specifications;
+package com.skwarek.onlineStore.data.model.product.specifications;
 
+import com.skwarek.onlineStore.data.entity.product.specifications.ProductSpecifications;
 import com.skwarek.onlineStore.data.entity.product.specifications.modules.*;
 
 /**
  * Created by Michal on 01.10.2016.
  */
-public class TabletSpecifications extends ProductSpecifications {
+public class NotebookSpecifications extends ProductSpecifications {
 
-    public TabletSpecifications() {
+    public NotebookSpecifications() {
         setCpu(new CPU());
         setGpu(new GPU());
         setRam(new RAM());
@@ -15,7 +16,7 @@ public class TabletSpecifications extends ProductSpecifications {
         setDisplay(new Display());
         setBattery(new Battery());
         setOs(new OS());
-        setCamera(new Camera());
+        setWebcam(new Webcam());
         setDimensions(new Dimensions());
         setWeight(new Weight());
     }
