+++ online-store/src/main/java/com/skwarek/onlineStore/data/model/product/specifications/PCSpecifications.java
@@ -1,21 +1,20 @@
-package com.skwarek.onlineStore.data.entity.product.specifications;
+package com.skwarek.onlineStore.data.model.product.specifications;
 
+import com.skwarek.onlineStore.data.entity.product.specifications.ProductSpecifications;
 import com.skwarek.onlineStore.data.entity.product.specifications.modules.*;
 
 /**
  * Created by Michal on 27.09.2016.
  */
-public class SmartphoneSpecifications extends ProductSpecifications {
+public class PCSpecifications extends ProductSpecifications {
 
-    public SmartphoneSpecifications() {
+    public PCSpecifications() {
         setCpu(new CPU());
         setGpu(new GPU());
         setRam(new RAM());
         setStorage(new Storage());
-        setDisplay(new Display());
-        setBattery(new Battery());
+        setPowerSupply(new PowerSupply());
         setOs(new OS());
-        setCamera(new Camera());
         setDimensions(new Dimensions());
         setWeight(new Weight());
     }
