+++ online-store/src/main/java/com/skwarek/onlineStore/web/controller/AdminProductController.java
@@ -5,7 +5,7 @@ import com.skwarek.onlineStore.data.entity.product.Manufacturer;
 import com.skwarek.onlineStore.data.entity.product.Product;
 import com.skwarek.onlineStore.data.entity.product.UploadFile;
 import com.skwarek.onlineStore.data.entity.product.specifications.ProductSpecifications;
-import com.skwarek.onlineStore.data.entity.product.specifications.SpecificationsFactory;
+import com.skwarek.onlineStore.data.model.product.specifications.SpecificationsFactory;
 import com.skwarek.onlineStore.service.*;
 import com.skwarek.onlineStore.web.editor.CategoryEditor;
 import com.skwarek.onlineStore.web.editor.ManufacturerEditor;
